requirements.txt
flask~=1.1.2
flask_login
flask_migrate
flask_wtf
flask_sqlalchemy
email_validator
gunicorn
pyjwt~=1.7.1
wtforms~=2.3.1
sqlalchemy~=1.3.18
jinja2~=2.11.2
requests~=2.24.0
flask_pymongo